saas-app/backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
channels>=4.0
channels-redis>=4.1
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.5
django-cors-headers>=4.3
